pyproject.toml
[project]
name = "django-10-12-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "factory-boy>=3.3.3",
    "faker>=38.2.0",
    "ipython>=9.8.0",
]
